[REF] product_attribute_custom_value_variant: Split the logic
Move the logic out of sale order model in order to simplify the calls
on that one.

Split also the tests with setup in a common class.
