+++ product_attribute_custom_value_variant/models/__init__.py
@@ -1,6 +1,6 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 
 from . import product_attribute_value
+from . import product_attribute_custom_value
 from . import product_attribute
 from . import product_product
-from . import sale_order_line
