+++ product_attribute_custom_value_variant/readme/CONTRIBUTORS.md
@@ -1,0 +1,2 @@
+- Simone Rubino - Aion Tech
+- Denis Roussel <denis.roussel@acsone.eu>
