+++ product_attribute_custom_value_variant/tests/__init__.py
@@ -1,3 +1,3 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 
-from . import test_sale
+from . import test_product
